src/icefabric_manage/pyproject.toml
[project]
name = "icefabric_manage"
version = "2025.5"
description = "All functionality for creating, updating, and deleting information from the icefabric catalog"
authors = [
  { name = "Tadd Bindas", email = "tadd.bindas@ertcorp.com" },
  { name = "Sylvia Chin", email = "sylvia.c.chin@rtx.com" },
  { name = "Daniel Cumpton", email = "dcumpton@rtx.com" },
  { name = "Quercus Hamlin", email = "qhamlin@asrcfederal.com" },
  { name = "Brock Hinkson", email = "brock.w.hinkson@rtx.com" },
]
maintainers = [
  { name = "Tadd Bindas", email = "tadd.bindas@ertcorp.com" },
  { name = "Sylvia Chin", email = "sylvia.c.chin@rtx.com" },
  { name = "Daniel Cumpton", email = "dcumpton@rtx.com" },
  { name = "Quercus Hamlin", email = "qhamlin@asrcfederal.com" },
  { name = "Brock Hinkson", email = "brock.w.hinkson@rtx.com" },
]
readme = "README.md"
dependencies = [
  "numpy==2.2.6",
  "pandas==2.2.3",
  "geopandas==1.0.1",
  "s3fs==2025.5.0",
  "pyarrow==20.0.0",
  "pyiceberg[sql-sqlite]==0.9.0",
  "python-dotenv==1.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
